#17442 - Generate idLote without using concurrency control

+++ NFe.Classes/Servicos/Autorizacao/enviNFe4.cs
@@ -7,7 +7,7 @@ namespace NFe.Classes.Servicos.Autorizacao
     [XmlRoot(ElementName = "enviNFe", Namespace = "http://www.portalfiscal.inf.br/nfe")]
     public class enviNFe4
     {
-        public enviNFe4(string versao, int idLote, IndicadorSincronizacao indSinc, List<NFe> nFe)
+        public enviNFe4(string versao, long idLote, IndicadorSincronizacao indSinc, List<NFe> nFe)
         {
             this.versao = versao;
             this.idLote = idLote;
@@ -28,7 +28,7 @@ namespace NFe.Classes.Servicos.Autorizacao
         /// <summary>
         ///     AP03 - Identificador de controle do envio do lote.
         /// </summary>
-        public int idLote { get; set; }
+        public long idLote { get; set; }
 
         /// <summary>
         ///     AP03a - Indicador de Sincronização
